requirements.txt
requests
openai
python-dotenv
schedule
gradio
loguru
markdown2
bs4
apscheduler
